changed generator creation function to not use incomplete batch

+++ main.py
@@ -130,7 +130,7 @@ def get_train_batch_generator(label_path_patterns, image_shape, batch_size):
     num_samples = len(label_paths)
     assert len(image_paths) == len(label_paths)
     # make number of samples so that all batches are same size
-    num_samples = math.floor(num_samples / batch_size) * batch_size
+    num_samples = int(math.floor(num_samples / batch_size) * batch_size)
 
     def get_batches_fn(batch_size):
         """
